Move get_s3_url to s3_url.py
